events api + tests

+++ tests/setup-tests.sh
@@ -1,15 +1,12 @@
 #!/bin/sh
-export $(grep -v '^#' envs/.env.test | xargs)
 [ -f .env ] && mv .env .env.backup
 cp envs/.env.test .env
 
-docker-compose -f docker-compose.test.yml -p gs-test up -d
-
 ./node_modules/.bin/ts-node tests/create-schema.ts 
+./node_modules/.bin/jest "$@"
 
-./node_modules/.bin/jest
-
-rm .env
-[ -f .env.backup ] && mv .env.backup .env
-
-docker-compose -p gs-test down
+if [ -z "$CI" ]
+then
+  rm .env
+  [ -f .env.backup ] && mv .env.backup .env
+fi
